+++ components/compositing/compositor_layer.rs
@@ -8,10 +8,11 @@ use euclid::length::Length;
 use euclid::point::{Point2D, TypedPoint2D};
 use euclid::rect::Rect;
 use euclid::size::TypedSize2D;
+use gfx_traits::{LayerId, LayerProperties, ScrollPolicy};
 use layers::color::Color;
 use layers::geometry::LayerPixel;
 use layers::layers::{Layer, LayerBufferSet};
-use msg::compositor_msg::{Epoch, LayerId, LayerProperties, ScrollPolicy};
+use msg::compositor_msg::Epoch;
 use msg::constellation_msg::{MouseEventType, PipelineId};
 use script_traits::CompositorEvent;
 use script_traits::CompositorEvent::{MouseButtonEvent, MouseMoveEvent};
