+++ components/compositing/compositor_task.rs
@@ -8,12 +8,12 @@ use CompositorMsg as ConstellationMsg;
 use compositor;
 use euclid::point::Point2D;
 use euclid::size::Size2D;
-use gfx_traits::PaintListener;
+use gfx_traits::{LayerId, LayerProperties, PaintListener};
 use headless;
 use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use layers::layers::{BufferRequest, LayerBufferSet};
 use layers::platform::surface::{NativeDisplay, NativeSurface};
-use msg::compositor_msg::{Epoch, FrameTreeId, LayerId, LayerProperties};
+use msg::compositor_msg::{Epoch, FrameTreeId};
 use msg::constellation_msg::{AnimationState, PipelineId};
 use msg::constellation_msg::{Image, Key, KeyModifiers, KeyState};
 use profile_traits::mem;
