+++ components/gfx/paint_context.rs
@@ -26,8 +26,7 @@ use euclid::side_offsets::SideOffsets2D;
 use euclid::size::Size2D;
 use filters;
 use font_context::FontContext;
-use gfx_traits::color;
-use msg::compositor_msg::LayerKind;
+use gfx_traits::{color, LayerKind};
 use net_traits::image::base::{Image, PixelFormat};
 use std::default::Default;
 use std::sync::Arc;
