+++ components/gfx/paint_task.rs
@@ -15,11 +15,11 @@ use euclid::rect::Rect;
 use euclid::size::Size2D;
 use font_cache_task::FontCacheTask;
 use font_context::FontContext;
-use gfx_traits::{color, PaintListener, PaintMsg as ConstellationMsg};
+use gfx_traits::{color, LayerId, LayerKind, LayerProperties, PaintListener, PaintMsg as ConstellationMsg, ScrollPolicy};
 use ipc_channel::ipc::IpcSender;
 use layers::layers::{BufferRequest, LayerBuffer, LayerBufferSet};
 use layers::platform::surface::{NativeDisplay, NativeSurface};
-use msg::compositor_msg::{Epoch, FrameTreeId, LayerId, LayerKind, LayerProperties, ScrollPolicy};
+use msg::compositor_msg::{Epoch, FrameTreeId};
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use paint_context::PaintContext;
 use profile_traits::mem::{self, ReportsChan};
