+++ components/layout/context.rs
@@ -14,9 +14,9 @@ use fnv::FnvHasher;
 use gfx::display_list::OpaqueNode;
 use gfx::font_cache_task::FontCacheTask;
 use gfx::font_context::FontContext;
+use gfx_traits::LayerId;
 use ipc_channel::ipc::{self, IpcSender};
 use msg::ParseErrorReporter;
-use msg::compositor_msg::LayerId;
 use net_traits::image::base::Image;
 use net_traits::image_cache_task::{ImageCacheChan, ImageCacheTask, ImageResponse, ImageState};
 use net_traits::image_cache_task::{UsePlaceholder};
