+++ components/layout/layout_task.rs
@@ -28,7 +28,7 @@ use gfx::font;
 use gfx::font_cache_task::FontCacheTask;
 use gfx::font_context;
 use gfx::paint_task::{LayoutToPaintMsg, PaintLayer};
-use gfx_traits::color;
+use gfx_traits::{color, LayerId, ScrollPolicy};
 use incremental::{LayoutDamageComputation, REFLOW, REFLOW_ENTIRE_DOCUMENT, REPAINT};
 use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use ipc_channel::router::ROUTER;
@@ -36,7 +36,7 @@ use layout_debug;
 use layout_traits::LayoutTaskFactory;
 use log;
 use msg::ParseErrorReporter;
-use msg::compositor_msg::{Epoch, LayerId, ScrollPolicy};
+use msg::compositor_msg::Epoch;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use net_traits::image_cache_task::{ImageCacheChan, ImageCacheResult, ImageCacheTask};
 use parallel::{self, WorkQueueData};
