+++ components/script/layout_interface.rs
@@ -10,10 +10,10 @@ use app_units::Au;
 use dom::node::LayoutData;
 use euclid::point::Point2D;
 use euclid::rect::Rect;
+use gfx_traits::LayerId;
 use ipc_channel::ipc::{IpcReceiver, IpcSender};
 use libc::uintptr_t;
 use msg::compositor_msg::Epoch;
-use msg::compositor_msg::LayerId;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use msg::constellation_msg::{WindowSizeData};
 use net_traits::image_cache_task::ImageCacheTask;
