+++ components/script/script_task.rs
@@ -42,6 +42,7 @@ use dom::window::{ReflowReason, ScriptHelpers, Window};
 use dom::worker::TrustedWorkerAddress;
 use euclid::Rect;
 use euclid::point::Point2D;
+use gfx_traits::LayerId;
 use hyper::header::{ContentType, HttpDate};
 use hyper::header::{Headers, LastModified};
 use hyper::method::Method;
@@ -61,7 +62,6 @@ use layout_interface::{ReflowQueryType};
 use layout_interface::{self, LayoutChan, NewLayoutTaskInfo, ReflowGoal, ScriptLayoutChan};
 use libc;
 use mem::heap_size_of_self_and_children;
-use msg::compositor_msg::LayerId;
 use msg::constellation_msg::{ConstellationChan, LoadData};
 use msg::constellation_msg::{MouseButton, MouseEventType, MozBrowserEvent, PipelineId};
 use msg::constellation_msg::{PipelineNamespace};
