+++ components/script_traits/lib.rs
@@ -14,6 +14,7 @@ extern crate app_units;
 extern crate canvas_traits;
 extern crate devtools_traits;
 extern crate euclid;
+extern crate gfx_traits;
 extern crate ipc_channel;
 extern crate libc;
 extern crate msg;
@@ -34,9 +35,10 @@ use euclid::Size2D;
 use euclid::length::Length;
 use euclid::point::Point2D;
 use euclid::rect::Rect;
+use gfx_traits::LayerId;
 use ipc_channel::ipc::{IpcReceiver, IpcSender};
 use libc::c_void;
-use msg::compositor_msg::{Epoch, LayerId};
+use msg::compositor_msg::Epoch;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId, WindowSizeData};
 use msg::constellation_msg::{Key, KeyModifiers, KeyState, LoadData, SubpageId};
 use msg::constellation_msg::{MouseButton, MouseEventType};
